+++ mycore-webtools/src/main/java/org/mycore/webtools/upload/MCRUploadResource.java
@@ -32,7 +32,6 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Objects;
 import java.util.function.Predicate;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -59,6 +58,7 @@ import org.mycore.datamodel.niofs.MCRPath;
 import org.mycore.datamodel.niofs.utils.MCRFileCollectingFileVisitor;
 import org.mycore.datamodel.niofs.utils.MCRTreeCopier;
 import org.mycore.frontend.fileupload.MCRPostUploadFileProcessor;
+import org.mycore.frontend.fileupload.MCRUploadHelper;
 import org.mycore.services.i18n.MCRTranslation;
 
 import jakarta.ws.rs.BadRequestException;
@@ -85,8 +85,6 @@ public class MCRUploadResource {
 
     @Context
     ContainerRequestContext request;
-
-    private Predicate<String> fileNameAllowedPredicate = null;
 
     private static List<MCRPostUploadFileProcessor> initProcessorList() {
         List<String> fileProcessorList = MCRConfiguration2.getString(FILE_PROCESSOR_PROPERTY)
@@ -147,19 +145,6 @@ public class MCRUploadResource {
         } catch (IOException e) {
             LOGGER.error("Could not set main file!", e);
         }
-    }
-
-    public Predicate<String> getFileNameAllowedPredicate() {
-        if (fileNameAllowedPredicate == null) {
-            initializeFileNameAllowedPredicate();
-        }
-        return fileNameAllowedPredicate;
-    }
-
-    private void initializeFileNameAllowedPredicate() {
-        String fileNamePattern = MCRConfiguration2.getStringOrThrow("MCR.FileUpload.FileNamePattern");
-        Pattern pattern = Pattern.compile(fileNamePattern);
-        fileNameAllowedPredicate = (str) -> pattern.matcher(str).matches();
     }
 
     @PUT
@@ -291,12 +276,10 @@ public class MCRUploadResource {
         @PathParam("objectID") String objectID,
         @QueryParam("size") String size) {
         String actualStringFileName = Paths.get(path).getFileName().toString();
-        if (!getFileNameAllowedPredicate().test(actualStringFileName)) {
-            String translation = MCRTranslation.translate("component.webtools.upload.invalid.fileName",
-                actualStringFileName);
-            return Response.status(Response.Status.BAD_REQUEST.getStatusCode(),
-                translation).entity(translation).build();
-
+        try {
+            MCRUploadHelper.checkPathName(actualStringFileName);
+        } catch (MCRException e){
+            return Response.status(Response.Status.BAD_REQUEST.getStatusCode(),e.getMessage()).build();
         }
 
         long sizeL = Long.parseLong(size);
@@ -337,9 +320,10 @@ public class MCRUploadResource {
         }
 
         String actualStringFileName = bucket.getRoot().relativize(filePath).getFileName().toString();
-        if (!getFileNameAllowedPredicate().test(actualStringFileName)) {
-            throw new ClientErrorException(actualStringFileName + " is not a valid file Name!",
-                Response.Status.BAD_REQUEST);
+        try {
+            MCRUploadHelper.checkPathName(actualStringFileName);
+        } catch (MCRException e){
+            throw new ClientErrorException(e.getMessage(), Response.Status.BAD_REQUEST.getStatusCode());
         }
 
         long maxSize = MCRConfiguration2.getOrThrow("MCR.FileUpload.MaxSize", Long::parseLong);
