+++ mycore-webtools/src/main/resources/components/webtools/config/messages_de.properties
@@ -11,7 +11,6 @@ component.session-listing.resolveHostNames  = IP-Adressen aufl\u00F6sen
 component.webtools.upload.title             = Datei-Upload
 component.webtools.upload.processing        = Das Hochladen der Dateien wird auf dem Server abgeschlossen.
 component.webtools.upload.error             = Beim Hochladen ist ein Fehler aufgetreten: 
-component.webtools.upload.invalid.fileName  = Ung\u00FCltiger Dateiname: {0}
 component.webtools.upload.invalid.fileSize  = Die Datei {0} ist zu gro\u00DF: {1} / {2}
 
 component.webtools.msie.learnMore=Mehr erfahren
