+++ mycore-webtools/src/main/resources/components/webtools/config/messages_en.properties
@@ -11,7 +11,6 @@ component.session-listing.resolveHostNames  = Resolve IP addresses
 component.webtools.upload.title             = File Upload
 component.webtools.upload.processing        = The file upload is processed on the server.
 component.webtools.upload.error             = Something went wrong with the Upload:
-component.webtools.upload.invalid.fileName  = Invalid Filename: {0}
 component.webtools.upload.invalid.fileSize  = The File {0} is to Big: {1} / {2}
 
 component.webtools.msie.learnMore=Learn more
